+++ crates/webidl-tests/main.rs
@@ -33,6 +33,7 @@ pub mod namespace;
 pub mod no_interface;
 pub mod promise;
 pub mod setlike;
+pub mod signature_stability;
 pub mod simple;
 pub mod throws;
 pub mod unstable;