+++ CHANGELOG.md
@@ -14,6 +14,11 @@
   APIs like the CSS Custom Highlight API which adds the `highlights` attribute to the `CSS` namespace.
   [#4930](https://github.com/wasm-bindgen/wasm-bindgen/issues/4930)
 
+* Added non-standard `mode` option for `FileSystemFileHandle.createSyncAccessHandle()`.
+  Also improved WebIDL generator to track stability at the signature level, allowing
+  stable methods to have unstable overloads.
+  [#4928](https://github.com/wasm-bindgen/wasm-bindgen/issues/4928)
+
 * Added a panic message when a getter as more than one argument. 
   [#4936](https://github.com/wasm-bindgen/wasm-bindgen/pull/4936)